update: restructured files
